# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Workforce OS v2 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artifacts/pending:
    get:
      operationId: listPendingArtifacts
      tags: [artifacts]
      summary: List pending review artifacts
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Paginated list of pending artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedArtifacts"
  /artifacts:
    get:
      operationId: listArtifacts
      tags: [artifacts]
      summary: List all artifacts with optional status filter
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_REVIEW, APPROVED, SENT, REJECTED, SUPPRESSED]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of outreach artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedArtifacts"
  /artifacts/{id}:
    get:
      operationId: getArtifact
      tags: [artifacts]
      summary: Get a single artifact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Artifact detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachArtifact"
        "404":
          description: Not found
  /artifacts/{id}/approve:
    post:
      operationId: approveArtifact
      tags: [artifacts]
      summary: Approve an outreach artifact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Approved artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachArtifact"
        "404":
          description: Not found
  /artifacts/{id}/reject:
    post:
      operationId: rejectArtifact
      tags: [artifacts]
      summary: Reject an outreach artifact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Rejected artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachArtifact"
  /artifacts/{id}/suppress:
    post:
      operationId: suppressArtifact
      tags: [artifacts]
      summary: Suppress recipient and artifact
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Suppressed artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachArtifact"
...
  /activity:
    get:
      operationId: getActivityStream
      tags: [activity]
      summary: Get agent activity stream
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, outbound, pipeline, conversations]
            default: all
      responses:
        "200":
          description: Activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with filters
      parameters:
        - name: stage
          in: query
          schema: { type: string }
        - name: minScore
          in: query
          schema: { type: integer }
        - name: intentSignal
          in: query
          schema: { type: string }
        - name: geo
          in: query
          schema: { type: string }
        - name: cohort
          in: query
          schema: { type: string }
        - name: industry
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, default: score_desc }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 25 }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedLeads"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Lead detail with evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
        "404":
          description: Not found
  /leads/{id}/trigger-outbound:
    post:
      operationId: triggerOutbound
      tags: [leads]
      summary: Trigger an outbound pipeline run for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "202":
          description: Pipeline run triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TriggerResult"
  /leads/bulk-suppress:
    post:
      operationId: bulkSuppressLeads
      tags: [leads]
      summary: Suppress multiple leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkLeadIds"
      responses:
        "200":
          description: Suppressed count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"
...
  /graph-runs/{id}/timeline:
    get:
      operationId: getGraphRunTimeline
      tags: [activity]
      summary: Get evidence timeline for a graph run
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Evidence timeline nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineNode"
components:
  schemas:
...
    Recipient:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        title: { type: string }
        company: { type: string }
        avatarUrl: { type: ["string", "null"] }
      required: [id, name, email, company]
    EvaluatorScores:
      type: object
      properties:
        pii: { type: number }
        hallucination: { type: number }
        citationCoverage: { type: number }
        toxicity: { type: number }
      required: [pii, hallucination, citationCoverage, toxicity]
    FactCitation:
      type: object
      properties:
        factId: { type: string }
        claim: { type: string }
        source: { type: string }
      required: [factId, claim, source]
    SendPolicy:
      type: object
      properties:
        liveSendEnabled: { type: boolean }
        postalAddressSet: { type: boolean }
        unsubscribeConfigured: { type: boolean }
        recipientSuppressed: { type: boolean }
      required: [liveSendEnabled, postalAddressSet, unsubscribeConfigured, recipientSuppressed]
    OutreachArtifact:
      type: object
      properties:
        id: { type: string }
        status:
          type: string
          enum: [PENDING_REVIEW, APPROVED, SENT, REJECTED, SUPPRESSED]
        recipient:
          $ref: "#/components/schemas/Recipient"
        subject: { type: string }
        bodyHtml: { type: string }
        citations:
          type: array
          items:
            $ref: "#/components/schemas/FactCitation"
        evaluatorScores:
          $ref: "#/components/schemas/EvaluatorScores"
        sendPolicy:
          $ref: "#/components/schemas/SendPolicy"
        createdAt: { type: string }
        approvedAt: { type: ["string", "null"] }
        sentAt: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        graphRunId: { type: ["string", "null"] }
        cohort: { type: ["string", "null"] }
      required: [id, status, recipient, subject, bodyHtml, citations, evaluatorScores, sendPolicy, createdAt]
    PaginatedArtifacts:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OutreachArtifact"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [items, total, page, limit]
    RejectInput:
      type: object
      properties:
        reason: { type: string }
      required: [reason]
...
    ActivityEvent:
      type: object
      properties:
        id: { type: string }
        agentName: { type: string }
        agentType:
          type: string
          enum: [sdr, content, ops, pipeline, reply, reporting]
        action: { type: string }
        stage: { type: string }
        timestamp: { type: string }
        artifactId: { type: ["string", "null"] }
        leadId: { type: ["string", "null"] }
      required: [id, agentName, agentType, action, stage, timestamp]
...
    IntentSignal:
      type: object
      properties:
        label: { type: string }
        confidence: { type: number }
      required: [label, confidence]
    Lead:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        title: { type: string }
        email: { type: string }
        company: { type: string }
        domain: { type: ["string", "null"] }
        companyLogoUrl: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        score: { type: integer }
        stage: { type: string }
        geo: { type: string }
        country: { type: ["string", "null"] }
        industry: { type: ["string", "null"] }
        headcountEstimate: { type: ["string", "null"] }
        cohort: { type: string, enum: [A, B] }
        emailStatus: { type: string, enum: [DELIVERABLE, HIGH_PROBABILITY, CATCH_ALL] }
        intentSignals:
          type: array
          items:
            $ref: "#/components/schemas/IntentSignal"
        lastContactedAt: { type: ["string", "null"] }
        sendPolicy:
          $ref: "#/components/schemas/SendPolicy"
        createdAt: { type: string }
      required: [id, name, email, company, score, stage, cohort, emailStatus, intentSignals, sendPolicy, createdAt]
    ScoreBreakdown:
      type: object
      properties:
        fit: { type: integer }
        intent: { type: integer }
        engagement: { type: integer }
        timing: { type: integer }
      required: [fit, intent, engagement, timing]
    EvidenceEventSummary:
      type: object
      properties:
        id: { type: string }
        eventType: { type: string }
        description: { type: string }
        timestamp: { type: string }
      required: [id, eventType, description, timestamp]
    LeadDetail:
      type: object
      properties:
        lead:
          $ref: "#/components/schemas/Lead"
        researchBrief: { type: string }
        scoreBreakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        recentEvidenceEvents:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceEventSummary"
      required: [lead, researchBrief, scoreBreakdown, recentEvidenceEvents]
    PaginatedLeads:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [items, total, page, limit]
    TriggerResult:
      type: object
      properties:
        runId: { type: string }
        queued: { type: boolean }
        message: { type: string }
      required: [runId, queued, message]
    BulkLeadIds:
      type: object
      properties:
        ids:
          type: array
          items: { type: string }
      required: [ids]
    BulkActionResult:
      type: object
      properties:
        affected: { type: integer }
      required: [affected]
...
    TimelineNode:
      type: object
      properties:
        id: { type: string }
        nodeType:
          type: string
          enum: [agent_run, llm_call, evaluator, tool_call, human_action]
        label: { type: string }
        summary: { type: string }
        reasoning: { type: ["string", "null"] }
        tokensUsed: { type: ["integer", "null"] }
        durationMs: { type: ["integer", "null"] }
        cost: { type: ["number", "null"] }
        score: { type: ["number", "null"] }
        timestamp: { type: string }
        children:
          type: array
          items:
            $ref: "#/components/schemas/TimelineNode"
      required: [id, nodeType, label, summary, timestamp, children]