# lib/api-zod/src/generated/types/outreachArtifact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Workforce OS v2 API
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluatorScores } from './evaluatorScores';
import type { FactCitation } from './factCitation';
import type { OutreachArtifactStatus } from './outreachArtifactStatus';
import type { Recipient } from './recipient';
import type { SendPolicy } from './sendPolicy';

export interface OutreachArtifact {
  id: string;
  status: OutreachArtifactStatus;
  recipient: Recipient;
  subject: string;
  bodyHtml: string;
  citations: FactCitation[];
  evaluatorScores: EvaluatorScores;
  sendPolicy: SendPolicy;
  createdAt: string;
  /** @nullable */
  approvedAt?: string | null;
  /** @nullable */
  sentAt?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  /** @nullable */
  graphRunId?: string | null;
  /** @nullable */
  cohort?: string | null;
}