# lib/api-zod/src/generated/types/timelineNode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Workforce OS v2 API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineNodeNodeType } from './timelineNodeNodeType';

export interface TimelineNode {
  id: string;
  nodeType: TimelineNodeNodeType;
  label: string;
  summary: string;
  /** @nullable */
  reasoning?: string | null;
  /** @nullable */
  tokensUsed?: number | null;
  /** @nullable */
  durationMs?: number | null;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  score?: number | null;
  timestamp: string;
  children: TimelineNode[];
}